kernel: Remove BlockInfo from kernel
The BlockInfo struct was moved to src/interfaces to src/kernel recently in
d3a479cb077d9a9a4451dc4ecb43fe0daf94f172 from #34079 which does not really make
sense because the BlockInfo struct is meant to be used by wallets and indexes
and is not referenced by kernel code.

The change moves the `BlockInfo` struct back from kernel to interfaces and
moves the `MakeBlockInfo` function from kernel to node. The `MakeBlockInfo`
function never belonged in the kernel was never used by it. It was added there
before the kernel library had taken shape, when it was unclear what the
node/kernel boundary would look like.
